Move template test to select_algorithm
